+++ tests/e2e/auth/signout/signout.test.ts
@@ -1,0 +1,13 @@
+import { expect } from "@playwright/test";
+import { test } from "../../homepage/home-page-fixture";
+
+test.describe("Sign Out Functionality", () => {
+    test("When the user clicks logout, then they are logged out and will thus be redirected to /signin", async ({
+        page,
+        homePage,
+    }) => {
+        await homePage.openLinkInUserMenuDialog("Sign Out", false);
+        await page.waitForURL("/signin");
+        await expect(page.getByRole("heading", { name: "Sign In" })).toBeVisible();
+    });
+});
