+++ tests/e2e/project/paper/paper-view-model.ts
