+++ tests/e2e/project/papers/project-papers-page-model.ts
