+++ tests/e2e/project/settings/general/project-name-settings.test.ts
@@ -1,0 +1,58 @@
+import { test } from "./project-general-settings-fixture";
+import { expect } from "@playwright/test";
+
+export let projectId: string = "";
+const projectName = "Project 1";
+
+test.describe("Renaming a project", () => {
+    /**
+     * Creates a new project before all tests in this test suite.
+     */
+    test.beforeAll(async ({ apiClient }) => {
+        await apiClient
+            .createProject({ name: projectName })
+            .then((project) => (projectId = project.response.id));
+    });
+
+    test.afterAll(async ({ apiClient }) => {
+        apiClient.softDeleteProject({ id: projectId });
+    });
+
+    test("When the user enters an invalid project name, then the name of the project remains unchanged.", async ({
+        page,
+        projectSettingsPage,
+    }) => {
+        await projectSettingsPage.changeProjectName("Project 1");
+        await projectSettingsPage.checkForErrors();
+        const projectNameHeader = page.getByRole("heading", { name: "Project 1" });
+        await expect(projectNameHeader).toBeVisible();
+        const toast = page.getByText("Please enter a new project name.");
+        await expect(toast).toBeVisible();
+    });
+
+    test("When the user enters an blank project name, then the name of the project remains unchanged.", async ({
+        page,
+        projectSettingsPage,
+    }) => {
+        await projectSettingsPage.changeProjectName(" ");
+        await projectSettingsPage.checkForErrors();
+        const projectNameHeader = page.getByRole("heading", { name: "Project 1" });
+        await expect(projectNameHeader).toBeVisible();
+        const toast = page.getByText("The project name cannot start or end with whitespace");
+        await expect(toast).toBeVisible();
+    });
+
+    test("When the user enters a valid project name, then the name of the project should be updated to this name.", async ({
+        page,
+        projectSettingsPage,
+    }) => {
+        await projectSettingsPage.changeProjectName("New Project");
+        await projectSettingsPage.checkForErrors();
+        const oldHeader = page.getByRole("heading", { name: "Project 1" });
+        await expect(oldHeader).not.toBeVisible();
+        const projectNameHeader = page.getByRole("heading", { name: "New Project" });
+        await expect(projectNameHeader).toBeVisible();
+        const toast = page.getByText("Successfully updated project name.");
+        await expect(toast).toBeVisible();
+    });
+});
