+++ tests/e2e/settings/shortcuts/shortcuts-setting-fixture.ts
@@ -1,0 +1,23 @@
+import { expect } from "@playwright/test";
+import { DevHomePage } from "../../homepage/home-page-model";
+import { DevShortcutsSettingsPage } from "./shortcuts-settings-model";
+import { test as base } from "../../utils/fixtures/shared-fixture";
+
+type ShortcutsSettingsPage = {
+    shortcutsSettingsPage: DevShortcutsSettingsPage;
+};
+
+/**
+ * Extends the default **custom** fixture by providing the page object models for the
+ * - shortcuts settings page
+ */
+export const test = base.extend<ShortcutsSettingsPage>({
+    shortcutsSettingsPage: async ({ page }, use) => {
+        await page.goto("/");
+        await expect(page.getByRole("heading", { name: "SnowballR" })).toBeVisible();
+
+        const homepage = new DevHomePage(page);
+        await homepage.openUserSettingInSidebar("Shortcuts");
+        await use(new DevShortcutsSettingsPage(page));
+    },
+});
