+++ tests/e2e/settings/shortcuts/shortcuts-settings-model.ts
